backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
pydantic-settings
cloudinary
python-multipart
fastapi-mail
